to_string replacement + better default error path implementation
