+++ conf_parsing/Config.hpp
@@ -6,7 +6,7 @@
 /*   By: julzibot <julzibot@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/23 19:15:02 by mstojilj          #+#    #+#             */
-/*   Updated: 2024/01/30 18:58:57 by julzibot         ###   ########.fr       */
+/*   Updated: 2024/01/31 08:35:17 by julzibot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,6 +30,7 @@ class LocationDir;
 typedef std::map<std::string, std::string> strstrMap;
 typedef std::map<std::string, std::map<std::string, LocationDir> > servLocMap;
 typedef std::map<std::string, std::map<std::string, strstrMap> > servInfos;
+std::string	to_string(int number);
 
 class LocationDir
 {
@@ -85,12 +86,12 @@ class Config
         std::string			get_cgi_type(std::string file_ext);
         strstrMap			get_hostMap(void) { return(this->hosts); };
         LocationDir			&getLocRef(std::string const &hostIP, int port, std::string route)
-							{ return (this->server_locs[hostIP + ":" + std::to_string(port)][route]); };
+							{ return (this->server_locs[hostIP + ":" + to_string(port)][route]); };
 		std::map<std::string, LocationDir> *getLocMap(std::string const &hostIP, int port)
-							{ return (&this->server_locs[hostIP + ":" + std::to_string(port)]); };
+							{ return (&this->server_locs[hostIP + ":" + to_string(port)]); };
 		bool				checkHostServExist(std::string const &tempHost, int const &port) {
 								for (servLocMap::iterator it = this->server_locs.begin(); it != this->server_locs.end(); it++)
-									if (it->first == tempHost + ":" + std::to_string(port))
+									if (it->first == tempHost + ":" + to_string(port))
 										return (true);
 									return (false);
 							}
