+++ server_files/jules.conf
@@ -77,7 +77,7 @@ http {
         # root   /Users/toshsharma/Documents/42cursus/Webserv/server_files/error_pages
         # error_pages  /Users/julzibot/Documents/42/Webserv/server_files/error_pages
         location / {
-            root   /Users/julzibot/Documents/42/Webserv/server_files
+            root   /Users/julzibot/Documents/42Webserv/server_files
             # root   /Users/toshsharma/Documents/42cursus/Webserv/server_files
             # root   /Users/julzibot/Documents/42/Webserv/server_files
             index  index.html
